chore: separate Reflection class from index.ts
